chore: improve types, add editor option props, notes from @cshaver
