+++ .yarnrc
@@ -1,5 +1,3 @@
 # THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
 # yarn lockfile v1
 
-
-yarn-path ".yarn/releases/yarn-1.18.0.js"
