+++ examples/monaco-graphql-webpack/src/index.html.ejs
@@ -27,7 +27,7 @@
       <div id="operation" style="height:70vh;"></div>
       <div id="variables" style="height:30vh;"></div>
     </div>
-    <div id="results" class="full-height column"></div>
+    <div id="results" class="full-height column" aria-label="Result Window" aria-live="polite" aria-atomic="true"></div>
   </div>
 </body>
