+++ packages/graphiql/src/api/actions/editorActions.ts
@@ -1,5 +1,3 @@
-import * as monaco from 'monaco-editor/esm/vs/editor/editor.api';
-
 export enum EditorActionTypes {
   EditorLoaded = 'EditorLoaded',
 }
