+++ packages/graphiql/src/api/common.ts
@@ -1,30 +1,5 @@
-import { buildClientSchema, IntrospectionQuery } from 'graphql';
-
-import {
-  introspectionQuery,
-  introspectionQueryName,
-} from '../utility/introspectionQueries';
-
 import { SchemaConfig, Fetcher } from '../types';
 import { GraphQLParams } from './types';
-import { observableToPromise } from '../utility/observableToPromise';
-
-export async function fetchSchema(fetcher: Fetcher) {
-  const rawResult = fetcher({
-    query: introspectionQuery,
-    operationName: introspectionQueryName,
-  });
-  const introspectionResponse = await observableToPromise(rawResult);
-  const parsedResult = JSON.parse(introspectionResponse);
-
-  if (!parsedResult || !('data' in parsedResult)) {
-    throw Error('error fetching introspection schema');
-  }
-
-  return buildClientSchema(parsedResult.data as IntrospectionQuery, {
-    assumeValid: true,
-  });
-}
 
 export function getDefaultFetcher(schemaConfig: SchemaConfig) {
   return async function defaultFetcher(graphqlParams: GraphQLParams) {
