+++ packages/graphiql/src/components/ResultViewer.tsx
@@ -7,13 +7,16 @@
 
 import React, { useEffect } from 'react';
 
+import type { EditorOptions } from '../types';
+
 import { useSessionContext } from '../api/providers/GraphiQLSessionProvider';
 import { useEditorsContext } from '../api/providers/GraphiQLEditorsProvider';
 
-import * as monaco from 'monaco-editor/esm/vs/editor/editor.api';
-
-type ResultViewerProps = {
+export type ResultViewerProps = {
   editorTheme?: string;
+  editorOptions?: EditorOptions;
+  onMouseUp?: (e: monaco.editor.IEditorMouseEvent) => void;
+  onRenderResults?: (e: monaco.editor.IModelChangedEvent) => void;
 };
 
 export function ResultViewer(props: ResultViewerProps) {
@@ -25,16 +28,19 @@ export function ResultViewer(props: ResultViewerProps) {
     // Lazily require to ensure requiring GraphiQL outside of a Browser context
     // does not produce an error.
 
-    viewerRef.current = monaco.editor.create(divRef.current as HTMLElement, {
-      value: session.results?.text ?? '',
-      readOnly: true,
-      language: 'json',
-      automaticLayout: true,
-      theme: props.editorTheme,
-    });
-    loadEditor('results', viewerRef.current);
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
+    const viewer = (viewerRef.current = monaco.editor.create(
+      divRef.current as HTMLElement,
+      {
+        value: session.results?.text ?? '',
+        readOnly: true,
+        language: 'json',
+        automaticLayout: true,
+        theme: props.editorTheme,
+      },
+    ));
+    loadEditor('results', viewer);
+    props.onMouseUp && viewer.onMouseUp(props.onMouseUp);
+    props.onRenderResults && viewer.onDidChangeModel(props.onRenderResults);
   }, []);
 
   useEffect(() => {
@@ -43,6 +49,15 @@ export function ResultViewer(props: ResultViewerProps) {
     }
   }, [session.results, session.results.text]);
 
+  React.useEffect(() => {
+    const editor = viewerRef.current;
+    if (!editor) {
+      return;
+    }
+    if (props.editorOptions) {
+      editor.updateOptions(props.editorOptions);
+    }
+  }, [props.editorOptions]);
   return (
     <section
       className="result-window"
