--- packages/graphiql/src/types.d.ts
@@ -1,6 +1,0 @@
-import { GraphQLType } from 'graphql';
-
-export namespace GraphiQL {
-  export type GetDefaultFieldNamesFn = (type: GraphQLType) => string[];
-  export type Maybe<T> = T | null | undefined;
-}
