+++ packages/graphiql/src/types/global.ts
@@ -1,0 +1,4 @@
+interface Window {
+  MonacoEnvironment: any;
+  GraphiQL: any;
+}
