+++ packages/graphiql/src/types/worker-loader.ts
@@ -1,0 +1,5 @@
+declare module 'worker-loader!*' {
+  export default class WebpackWorker extends Worker {
+    constructor();
+  }
+}
