+++ packages/monaco-graphql/package.json
@@ -17,12 +17,13 @@
   ],
   "dependencies": {
     "graphql-language-service-utils": "^2.4.0-alpha.3",
-    "graphql-languageservice": "^2.4.0-alpha.5",
-    "monaco-editor-core": "^0.20.0",
-    "monaco-languages": "^1.10.0",
-    "vscode-languageserver-textdocument": "^1.0.1"
+    "graphql-languageservice": "^2.4.0-alpha.5"
   },
   "devDependencies": {
     "vscode-languageserver-types": "3.15.1"
+  },
+  "peerDependencies": {
+    "monaco-editor": "^0.20.0",
+    "monaco-languages": "^1.10.0"
   }
 }
